// models/Obligation.ts
export class Obligation {
    username: String = "";
    firm: String = "";
    startDate: Date = new Date();
    squareFootage: Number = 0;
    privatePublicGarden: Number = 0;
    poolFootage: Number = 0;
    greenFootage: Number = 0;
    sittingFootage: Number = 0;
    fountainFootage: Number = 0;
    sittingNumber: Number = 0;
    description: String = "";
    services: String = "";
    accepted: Boolean = false;
    endDate: Date = new Date();
    owner: String = "";
    made: Date = new Date();
    poolNumber: number = 0;
    fountainNumber: number = 0;
    comment: string = "";
}