// models/Firm.ts
export class Firm {
    name: String = "";
    address: String = "";
    services: String = "";
    prices: String = "";
    location: String = "";
    contact: String = "";
    startPause: Date = new Date();
    endPause: Date = new Date();
    ratings: Number = 0;
    numberOfRates: Number = 0;
}