// models/Decorater.ts
export class Decorater {
    username: String = "";
    password: String = "";
    firstname: String = "";
    lastname: String = "";
    gender: Number = 0;
    address: String = "";
    phoneNumber: String = "";
    email: String = "";
    picture: String = "";
    creditCardNumber: String = "";
    type: Number = 0;
    activated: boolean = false;
    blocked: boolean = false;
    busyUntil: Date = new Date(1, 0, 1);
    busyFrom: Date = new Date(1, 0, 1);
    firm: String = "";
}